4-Codigo actualizado con todos los  metodos get, post, put y delete

+++ routes/api/customers.js
@@ -10,22 +10,32 @@ router.get('/', async (req, res) => {
     res.json(customers);
 });
 
-
 router.get('/:customerId', async (req, res) => {
-    const customers = await Customer.findAll({
-        where : { id: req.params.customerId}   
-       });  
+    const Id= await Customer.findOne({where: {id: req.params.customerId}});
+    if (Id) {  
 
-    res.json( customers);
+        const customers = await Customer.findAll({
+        where : { id: req.params.customerId}   
+        });  
+
+        res.json(customers);
+
+    }else{
+
+        res.json({ error: 'Id no existe en la bd'});
+    } 
 });
 
  
 router.post('/',  [
     check('nit','El Nit del Cliente es obligatorio').not().isEmpty(),
     check('verification_digit','El Dígito de Verificación es obligatorio').not().isEmpty(),
+    check('verification_digit','El Dígito de Verificación debe ser numérico').isNumeric(),
+    check('verification_digit','El Dígito de Verificación debe ser de un solo caracter').isLength({ max: 1}),
     check('customer_name','El Nombre de Cliente es obligatorio').not().isEmpty(),
     check('address','La Address es obligatoria').not().isEmpty(),
     check('phone','El Phone es obligatorio').not().isEmpty(),
+    check('phone','El Phone debe ser numérico').isNumeric(),
     check('email','El Email es Incorrecto').isEmail(),
     check('contact_name','El Nombre de Contacto es obligatorio').not().isEmpty()
 ], async (req, res) => {
@@ -34,27 +44,68 @@ router.post('/',  [
     if (!errors.isEmpty()) {
         return res.status(422).json({errores: errors.array()})
     }  
+
+    const Nit = req.body.nit;
+    if(Nit > 2147483647) {
+        res.json({ error: 'Desbordamiento de pila'});
+        
+    }else{
+       const nit = await Customer.findOne({where: {nit: req.body.nit}});
+        if (!nit) {
+            const customer = await Customer.create(req.body);
+            res.json(customer);
+        }
+        else
+        {
+            res.json({ error: 'Nit ya existe en la bd'});
+        }      
+    }
     
-    const customer = await Customer.create(req.body);
-    res.json(customer);
 });
 
 
 router.put('/:customerId', async (req, res) => {
-    const customer = await Customer.update(req.body, {
-        where : { id: req.params.customerId}   
-       });
-    res.json( {success: 'Updated Customer'} );
+
+    const Id= await Customer.findOne({where: {id: req.params.customerId}});
+    if (Id) {     
+            const customer = await Customer.update(req.body, {
+            where : { id: req.params.customerId}   
+            });
+            res.json( {success: 'Updated Customer'} );
+    }
+    else
+    {
+        res.json({ error: 'Id no existe en la bd'});
+    } 
+
 });
 
 
-router.delete('/:customerId', async (req, res) => {
-    await Customer.destroy({
-        where : { id: req.params.customerId}   
-       });  
+router.delete('/:customerId',async (req, res) => {   
+const identification = req.params.customerId;    
+if(identification)
+{
+    const Id= await Customer.findOne({where: {id: req.params.customerId}});
+    if (Id) {     
+        await Customer.destroy({
+            where : { id: req.params.customerId}   
+           });  
+    
+        res.json( {success: 'Deleted Customer'} );
+    }
+    else
+    {
+        res.json({ error: 'Id no existe en la bd'});
+    } 
 
-    res.json( {success: 'Deleted Customer'} );
+}else 
+{
+res.status(404).send("Sorry, we cannot find that!"); 
+//console.log("Trabajando");
+ //res.json({ error: 'Id no puede ser vacio'});
+ //res.json( {success: 'Id no puede ser vacio'} );
+}
+
 });
-
 
 module.exports = router;
